+++ neo/framework/DeclParticle.cpp
@@ -409,6 +409,13 @@ idParticleStage *idDeclParticle::ParseParticleStage( idLexer &src ) {
 			stage->gravity = src.ParseFloat();
 			continue;
 		}
+		if ( !token.Icmp( "softeningRadius" ) ) { // #3878 soft particles
+			common->Warning( "Particle %s from %s has stage with \"softeningRadius\" attribute, which is currently ignored (we soften all suitable particles)\n",
+					this->GetName(), src.GetFileName() );
+			// DG: disable this for now to avoid breaking the game ABI
+			//stage->softeningRadius = src.ParseFloat();
+			continue;
+		}
 
 		src.Error( "unknown token %s\n", token.c_str() );
 	}
@@ -733,6 +740,9 @@ idParticleStage::idParticleStage( void ) {
 	hidden = false;
 	boundsExpansion = 0.0f;
 	bounds.Clear();
+	// DG: disable softeningRadius for now to avoid breaking the game ABI
+	//     (will always behave like if softeningRadius = -2.0f)
+	//softeningRadius = -2.0f;	// -2 means "auto" - #3878 soft particles
 }
 
 /*
@@ -806,6 +816,8 @@ void idParticleStage::Default() {
 	randomDistribution = true;
 	entityColor = false;
 	cycleMsec = ( particleLife + deadTime ) * 1000;
+	// DG: disable softeningRadius for now to avoid breaking game ABI
+	//softeningRadius = -2.0f;	// -2 means "auto" - #3878 soft particles
 }
 
 /*