+++ neo/renderer/Material.cpp
@@ -1434,6 +1434,10 @@ void idMaterial::ParseStage( idLexer &src, const textureRepeat_t trpDefault ) {
 			ss->drawStateBits |= GLS_DEPTHMASK;
 			continue;
 		}
+		if ( !token.Icmp( "ignoreDepth" ) ) { // Added in #3877.
+			ss->drawStateBits |= GLS_DEPTHFUNC_ALWAYS;
+			continue;
+		}
 		if ( !token.Icmp( "alphaTest" ) ) {
 			ss->hasAlphaTest = true;
 			ss->alphaTestRegister = ParseExpression( src );