+++ neo/renderer/tr_local.h
@@ -109,6 +109,7 @@ SURFACES
 
 // drawSurf_t are always allocated and freed every frame, they are never cached
 static const int	DSF_VIEW_INSIDE_SHADOW	= 1;
+static const int	DSF_SOFT_PARTICLE		= 2; // #3878 - soft particles
 
 typedef struct drawSurf_s {
 	const srfTriangles_t	*geo;
@@ -121,6 +122,7 @@ typedef struct drawSurf_s {
 	int						dsFlags;			// DSF_VIEW_INSIDE_SHADOW, etc
 	struct vertCache_s		*dynamicTexCoords;	// float * in vertex cache memory
 	// specular directions for non vertex program cards, skybox texcoords, etc
+	float					particle_radius;	// The radius of individual quads for soft particles #3878
 } drawSurf_t;
 
 
@@ -981,6 +983,10 @@ extern idCVar r_materialOverride;		// override all materials
 
 extern idCVar r_debugRenderToTexture;
 
+extern idCVar r_glDebugContext; // DG: use debug context to call logging callbacks on GL errors
+extern idCVar r_enableDepthCapture; // DG: disable capturing depth buffer, used for soft particles
+extern idCVar r_useSoftParticles;
+
 /*
 ====================================================================
 
@@ -1179,7 +1185,7 @@ viewEntity_t *R_SetEntityDefViewEntity( idRenderEntityLocal *def );
 viewLight_t *R_SetLightDefViewLight( idRenderLightLocal *def );
 
 void R_AddDrawSurf( const srfTriangles_t *tri, const viewEntity_t *space, const renderEntity_t *renderEntity,
-					const idMaterial *shader, const idScreenRect &scissor );
+					const idMaterial *shader, const idScreenRect &scissor, const float soft_particle_radius = -1.0f ); // soft particles in #3878
 
 void R_LinkLightSurf( const drawSurf_t **link, const srfTriangles_t *tri, const viewEntity_t *space,
 				   const idRenderLightLocal *light, const idMaterial *shader, const idScreenRect &scissor, bool viewInsideShadow );
@@ -1315,6 +1321,10 @@ typedef enum {
 	FPROG_AMBIENT,
 	VPROG_GLASSWARP,
 	FPROG_GLASSWARP,
+	// SteveL #3878: soft particles
+	VPROG_SOFT_PARTICLE,
+	FPROG_SOFT_PARTICLE,
+	//
 	PROG_USER
 } program_t;
 
@@ -1362,9 +1372,22 @@ typedef enum {
 	PP_SPECULAR_MATRIX_S,
 	PP_SPECULAR_MATRIX_T,
 	PP_COLOR_MODULATE,
-	PP_COLOR_ADD,
+	PP_COLOR_ADD, // 17
 
-	PP_LIGHT_FALLOFF_TQ = 20	// only for NV programs
+	PP_LIGHT_FALLOFF_TQ = 20,	// only for NV programs - DG: unused
+	PP_GAMMA_BRIGHTNESS = 21, // DG: for gamma in shader: { r_brightness, r_brightness, r_brightness, 1/r_gamma }
+	// DG: for soft particles from TDM: reciprocal of _currentDepth size.
+	//     Lets us convert a screen position to a texcoord in _currentDepth
+	PP_CURDEPTH_RECIPR  = 22,
+	// DG: for soft particles from TDM: particle radius, given as { radius, 1/(fadeRange), 1/radius }
+	//     fadeRange is the particle diameter for alpha blends (like smoke), but the particle radius for additive
+	//     blends (light glares), because additive effects work differently. Fog is half as apparent when a wall
+	//     is in the middle of it. Light glares lose no visibility when they have something to reflect off.
+	PP_PARTICLE_RADIUS  = 23,
+	// DG: for soft particles from TDM: color channel mask.
+	//     Particles with additive blend need their RGB channels modifying to blend them out
+	//     Particles with an alpha blend need their alpha channel modifying.
+	PP_PARTICLE_COLCHAN_MASK = 24,
 } programParameter_t;